Adding lambdasWillNotWork cause pointer couldn't be related
associative containers uses std::less<key_type> for comparison because
of that we can define a set of pointers or use a pointer as the key in a map.

Fact is pointers are object (wouldn't say address but have it's own memory) so we can use them as a key.